+++ CHANGELOG.md
@@ -34,7 +34,7 @@
 
 ## [0.3.0]
 
-* Provide previous value to builder
+* Provide previous value & keys to builder
 * Make nested `fragment()` calls act like React's `useMemo` 
 * Fix nested fragments support
 * Remove named fragments support